test: explicitly cast time() to int for the seed
(hopefully) silences an analyzer warning:

test/compose.c:955:16: store_truncates_time_t: A "time_t" value is stored
in an integer with too few bits to accommodate it. The expression
"time(NULL)" is cast to "int".

+++ test/compose.c
@@ -952,7 +952,7 @@ main(int argc, char *argv[])
     if (argc == 2) {
         seed = atoi(argv[1]);
     } else {
-        seed = time(NULL);
+        seed = (int)time(NULL);
     }
     fprintf(stderr, "Seed for the pseudo-random generator: %d\n", seed);
     srand(seed);
